requirements.txt
requests
python-dotenv
argparse
pytest
patreon